Add verification example using a setup plcore file
Apparently the prover first executes the second file argument,
and then any variables in the spec with names like ENV are filled
with the contents of the corresponding cell in the final configuration
from that execution, before the proof is attempted.

This can be verified by testing various modifications to the new
test - renaming ENV uniformly to E2 makes the proof fail,
as does changing the module name in int-addition-lib.plcore so
its definition does not end up with the expected qualified name.