translation: @init -> @allocInit; Generate alloc library in <iele> cell
We rename @init to @allocInit since this is not needed at contract initialization,
but rather should be run before alloc is called. Right now it is not called
at all (as it was before this commit).

Generating the functions directly into the <iele> cell simplifies the code
and is the first step towards reducing the number of auxilary cells needed to
generate the dispatch table.