plutus-core: Make `BoundedInt`s terms, use rules over Terms instead of Constants
Previously, the "!" syntax claimed to be a function from Ints to Constant.
However error terms are not infact `Constant`s. This change fixes that.

I'd like to specify some sort of reachability claim (that should be provable using
case analysis) that would show that every (con S ! I) reduces to either
a BoundedInt term or an Error term. Currently that spec is commented