Adding examples and output generated by UPLC (#60)
* Adding examples and output generated by UPLC

The uplc program contains several Plutus examples along with an
evaluator for Untypted Plutus. Although the examples are generated with
some notion of random generation, it's important for us to have a set of
static tests for repeatability.

Examples were generated using the uplc-example-gen.sh script.

* enable uplc-examples tests on Makefiles and CI

For now, set all tests to failing. As we develop the semantics, remove
tests from the tests/failing file.

* Address codereview feedback

Mostly stylistic changes