Enable conformance testing (#69)
* Enable conformance testing

Create a new target to extract the kplc output and compare it to the
expected output generated by uplc. Since the semantics implements
support for simple programs, enable several tests.

* kplc: add --results-only option

Keeping the k cell extraction in the Makefile confines this
Move k cell extraction from Makefile to kplc as a new option called
--results-only. This is useful for comparing output between kplc and
uplc program evaluation output.

* kplc: move result_only logic to run_interpreter

The --results-only option only makes sense for krun so move this from
the more general-purpose notif_exec to run_interpreter.

* kplc: replace redundant command with a conditional command

* Makefile: pass test option variables instead of creating a new target

* Update Makefile

* Update kplc

* kplc: move variable definition lower in the file

The result_only variable is specific to a command rather than the whole
file. Move it to the section where options are handled.

Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>