Confromance test: evaluate test cases using the `--print-mode Classic` option (#70)
* scripts: run `uplc evaluate` with --print-mode Classic

The default version is called `Debug`. This mode annotates variable
names with an `_x` where x is a number. Since this is likely used
for a debugging scenario and makes test output comparison difficult
between uplc and kplc, use the Classic mode which retains the original
variable names.

* Conformace tests: update tests using output from script

* Enable several more tests

The newly generated tests allow a few tests to pass. Enable them to run.