Makefile: compare test results against kplc and uplc
When running the tests, the command is run twice: once to compare the
output from kplc to uplc and a second time to compare the output from
kplc to an older output of kplc. The former ensures that the semantics
matches `uplc` output and the latter ensures that we are testing values
contained in cells other than the `<k>` cell.

Also, add change update-results-with-uplc to update-results. The new
update-results target updates all non-failing test output to include the
output from running `uplc evaluate` to the `.expected` file and saves the
output from running `kplc run` to the `.krun.expected`
file.