Clean up Makefile and document tests (#83)
* Makefile: Add target for conformance tests

This target is intended to run all tests that are executed by krun to
ensure that the semantics conform to the result of uplc.

* README: update test documentation

* add conformance-test target as PHONY

* tests: run all existing tests as conformance tests

Since all existing tests simply evaluate the input program, the output
of these tests should be compared to the uplc output. Since `uplc
evaluate` only emits the result of the evaluation, use the --result-only
option to run all tests.

* Makefile: add target "update-results-using-uplc"

When adding new tests that involve program evaluation, we need to match
the evaluation of the program against other implementations. This new
target will run the tests using uplc and store the results in the
expecte output file.

* Update expected results using output from uplc

* Makefile: fix variable assignment for target and update PHONY

In order to set the values of multiple variables for a make target, it
needs to be split into multiple targets rather than a recipe containing
exports.

* Makefile: compare test results against kplc and uplc

When running the tests, the command is run twice: once to compare the
output from kplc to uplc and a second time to compare the output from
kplc to an older output of kplc. The former ensures that the semantics
matches `uplc` output and the latter ensures that we are testing values
contained in cells other than the `<k>` cell.

Also, add change update-results-with-uplc to update-results. The new
update-results target updates all non-failing test output to include the
output from running `uplc evaluate` to the `.expected` file and saves the
output from running `kplc run` to the `.krun.expected`
file.

* tests: update expected test results

This was done by runnning the command `make update-results`

* README: update documentation

* adding expected output generated from kplc run