Move syntax declarations and use MarkDown
The top of the concrete syntax module started with a list of builtin
functions. Although builtin functions are important part of the
language, this is not as important as the actual syntax. Move the more
interesting syntax declarations above the builtin functions. Also,
replace comments that take up entire lines with markdown.

There is no functional change intended with this change.