Syntax: add experimental syntax for data
In the syntax of the constant value `(con data #...)` the ... represents
a bytestring of the CBOR encoding of the following haskell ADT:

Data data =
      Constr Integer [Data]
    | Map [(Data, Data)]
    | List [Data]
    | I Integer
    | B BS.ByteString

Add an alternative syntax for a data constant that is easier for humans
to read. For now, the syntax is modelled after haskell syntax. However,
this is experimental and is subject to change.