syntax: add initial support for flat-format
The flat format is a binary encoding. Since K doesn't parse raw data
files, the input will be the result of the hex-dump of the input file by
running commands such as `xxd -ps`. Since we want to avoid token
conflicts between the binary data and the `Int` sort, the input format
will start with a `#` character followed by hex digits. The `#`
character serves to disambiguate between the input and `Int`.