Add initial support for flat format parser (#111)
* syntax: add initial support for flat-format

The flat format is a binary encoding. Since K doesn't parse raw data
files, the input will be the result of the hex-dump of the input file by
running commands such as `xxd -ps`. Since we want to avoid token
conflicts between the binary data and the `Int` sort, the input format
will start with a `#` character followed by hex digits. The `#`
character serves to disambiguate between the input and `Int`.

* kplc: fix indentation, no functional change

* kplc: add new option --flat-format to `kplc run`

This option assumes that the input is encoded in the flat format (a
binary format described in the plutus specification). Since the input is
binary, this command will use `xxd` to dump the hex format of the input
before running `krun`.

* Add test-flat target and simple test cases

`make test-flat` will run the new tests against a temporary expected
output. This serves as a starting point to develop the flat format
parser using small examples to begin.

* docker: add xxd as a dependency