Makefile: run tests only when the expected output exists
In some cases, expected output may not exist due to support in the
`uplc` implementation or `kplc` implementation. At this time, textual
syntax for `list`, `pair, and `data` constants do not exist and we plan
on addressing this in the future. For now, these tests will only be run
against existing kplc output and the comparison of uplc will be enabled
when the tool supports the textual syntax.