Add syntax for pair and list (#101)
* Add syntax for pair and list

* Add simple programs for lists and pairs

* Syntax: add experimental syntax for data

In the syntax of the constant value `(con data #...)` the ... represents
a bytestring of the CBOR encoding of the following haskell ADT:

Data data =
      Constr Integer [Data]
    | Map [(Data, Data)]
    | List [Data]
    | I Integer
    | B BS.ByteString

Add an alternative syntax for a data constant that is easier for humans
to read. For now, the syntax is modelled after haskell syntax. However,
this is experimental and is subject to change.

* Add tests for experimental textual data

* fix listOfList example with approperiate type

* Makefile: run tests only when the expected output exists

In some cases, expected output may not exist due to support in the
`uplc` implementation or `kplc` implementation. At this time, textual
syntax for `list`, `pair, and `data` constants do not exist and we plan
on addressing this in the future. For now, these tests will only be run
against existing kplc output and the comparison of uplc will be enabled
when the tool supports the textual syntax.

* Add expected ouptuts for new textual syntax