This is commit is the first in a series of commits to adapt KPlutus to the CEK machine as specified in the draft of Ja. 27th, 2022.
It introduces changes to UPLC-SYNTAX modules. Perhaps the most
noteworthy ones reggard the change on the Value sort and the removal
of the abstract syntax.

A Value now is a closure of a constant, a lam or a builtin together
with its declaring environment. In the case of builtins, it also has
the list of its arguments, that grows as the arguments are being
evaluated in the context of an iterated application. This new form of
representing builtins allows to completely remove the abstract syntax
we used before to rewrite builtins.