Module UPLC-ENVIRONMENT
This commit adds module `UPLC-ENVIRONMENT` that implements an
environment which is a stack of pairs of identifier and value. Each
pair is called a binding. It also provide some standard functions on
environments: #is that checks if a binding is in the environment,
that inserts a binding at the top of the environment.