CEK machine of the draft spec. of Jan 27th, 2022 (#135)
* This is commit is the first in a series of commits to adapt KPlutus to
the CEK machine as specified in the draft of Ja. 27th, 2022.

It introduces changes to UPLC-SYNTAX modules. Perhaps the most
noteworthy ones reggard the change on the Value sort and the removal
of the abstract syntax.

A Value now is a closure of a constant, a lam or a builtin together
with its declaring environment. In the case of builtins, it also has
the list of its arguments, that grows as the arguments are being
evaluated in the context of an iterated application. This new form of
representing builtins allows to completely remove the abstract syntax
we used before to rewrite builtins.

* Module UPLC-ENVIRONMENT

This commit adds module `UPLC-ENVIRONMENT` that implements an
environment which is a stack of pairs of identifier and value. Each
pair is called a binding. It also provide some standard functions on
environments: #is that checks if a binding is in the environment,
that inserts a binding at the top of the environment.

* UPLC-CONFIGURATION

- Removed the sort for closures.

- Changed the name from ATerm to Frame to cope with the draft
  specification.

- Changed how terms of sort Frame (previously ATerm) are built as now
  a left application takes a term and an environment and a right
  application involves only a value.

- The env cell is now of type Env.

* Module UPLC-SEMANTICS

- Most of the rules changed as a result of closing values with the
  environment. The difference between the notation in the draft and in
  the implementation in minimal.

- The last rule is a general rule reposible for gathering the
  arguments of a builtin. Each builtin specifies the number of
  arguments it has and what to do when all arguments are
  evaluated. The last case happes when the argument counter reaches 0.

* Builtin modules

This commit adds all builtin modules. Even though there are many
changes, they follow the same pattern. Each builtin must state how
many arguments it has and which function should be called when all
arguments are gathered.

* Makefile and tests/failing

- Values are now between angled brackets and they differ from `uplc`
  output because of that. A proper pretty printing should be written
  to accomodate this change but meanwhile we use sed to make the
  output fit.

- With the new semantics for values and application, problematic
  programs in tests/uplc-examples are now passing the test
  harness. There are however three files that still do not pass. File
  tests/failing document who they are and why they fail.

* Test files

This commit updates the test base with expected files.