Update Flat Parser Using BitStream Sort (#136)
* Add a new `BitStream` sort to parse flat format

The flat format is not aligned on a byte boundary. This sort faciliates
indexing a `Bytes` sort by bit. Add the file bitstream.md that contains
the sort definition along with several useful function. This will be
used in parsing flat formats in later changes.

* Update flat parser using BitStream sort

The previous implementation used the `Bytes` sort to implement a very
limited prototype of the flat parser. The main difficulty was that the
`Bytes` sort only supported byte-wise indexing. Since the flat format
is not at all aligned to a byte (except at the end of a term
declaration), changing the parser to index by bit made the code much
more suitable to parse this format. This change enables parsing for the
current flat format tests. In the near future, this code needs to be
refactored to be able to parse nested terms and variable-length
constants.

* update expected output of flat format tests

Update the expected output using the new flat parser.

* trivial: fix comment style, no functional change