Fix variable identifier syntax (#139)
* Fix syntax for variable identifier

The syntax for variable names is different than the `Id` sort defined in
`domains.md`. Most notably, it contains a `'` as an allowed character.
Replace the use of `Id` with a new `UplcId` that matches the
specification.

* Enable test cases with variables that contain `'`

NatRoundTrip and ScottListSum were disabled because they have variables
names with `'`. Since the identifiers were updated to match the spec,
allow these tests to run. Note: factorial.uplc fails due to an unknown
reason that needs investigation.