Fixed a bug on builtin evaluation.
Even though this is a large PR, in terms of the number of modified
lines of code, the modification is simple with respect to
arguments it expects in order to actually call the proper builtin
function that does the job. This decrementing was being computed all
the way down to 0. It led however to states such as V ~> [ < builtin
bn L 0 > _], where V is a Value, bn is a builtin name, and L is a
(semantic) list with the arguments of bn. There are no such rule to
further reduce such states. What needs to be done is to let evaluation
go down to 1 and then directly call the function that does the job,
where states as above are not reachable.

So this PR essentially changes builtin evaluation to

  rule <k> V:Value ~> [ < builtin BN:BuiltinName L:List I:Int > _] =>
           < builtin BN (L ListItem(V)) (I -Int 1) > ... </k>
  requires I >Int 1

  rule <k> V:Value ~> [ < builtin BN:BuiltinName L:List 1 > _] =>
           #eval(BN, (L ListItem(V))) ... </k>

and adapts every builtin to cope with them, which means essentially to
implement the function that does the job as an #eval function.