Enable K unit testing (#143)
* clean up makefile

Remove unnecessary targets like foo, add correct dependencies to the
build-kplutus target and add all relevant files as dependencies to this
target so that the rebuilds are triggered properly.

* replace --backend-dir with --directory

`--backend-dir` points to the directory with the definition. In other
words, it serves the same purpose as the --directory option for
`kompile` and `kprovex`. Replace `--backend-dir` with `--directory` to
assimilate the runner script with options provided by the kframework
utilities.

* kplc: add kplc prove command

This command runs kprovex on the input file to execute symbolic tests and proofs.

* Enable proof tests in Makefile and CI

This adds a new target in the Makefile called test-prove and run all
proof tests in the `unit-tests` directory.

* Add a simple proof example

* Makefile: generalize test-prove harness

Generalize the test-prove harness to execute all claims in the
unit-tests directory.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>