Data builtins (#144)
* Add builtin `constrData`

This commit adds support to builtin `constrData`. It takes an integer
and a list of data and returns a constant data built with data constructor
`Constr`.

* Add `mapData`

This commit adds support for builtin `mapData`. As with `constrData`,
the only detail here is that we have to "coerce" between
`ConstantList` and `DataPairList`. This is done by function
`mkDataPairList` in `uplc-data-builtins.md`.

* Add `listData`

This commit adds support to `listData` builtin. Again, the only detail
is to create a `DataList` from a `ContantList`.

* Add `iData`

Builtin `iData` constructs a data term from an integer.

* Add `bData`

Builtin `bData` creates a data constant from a bytestring.

* Add `unConstrData`

Given a data constant built with constructor `Constr`, returns a pair
whose projections are the arguments of `Constr`.

* Add `unMap`

Given a data constant constructed with `Map` returns a constant of
list of pairs of data.

* Add `unListData`

Given a data contant built as a list, returns a list of data.

* Add `unIData`

Builtin `unIData` returns an integer constant from a data constant
representing an integer.

* Add `unBData`

Given a data contant representing a bytestring, returns the
asssociated bytestring.

* Add `equalsData`

This builtin takes two data constants and compares them.

* Add `mkPairData`

This builtin generates a pair of data from two given data constants.

* Add `mkNilData`

Returns an empty list of type list of data.

* Add `mkNilPairData`

Returns an empty list of type list of pair of data.

* Add error rules

This last commit add error rules for all data builtins. It also moves
`chooseData` builtin from `uplc-polymorphic-builtins.md` to
`uplc-data-builtins.md`.

* Updated Makefile

File `uplc-data-builtins.md` was missing from `kplutus-files` list.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>