flat parser: return only the term when reaching a leaf
There is no need to return the BitStream sort when reaching a leaf term.
Returning the BitStream term on every rewrite creates difficulty in
reducing the #readProgramTerm to an actual term when the programs contain
nested terms such as

`(delay (con unit ())`

At this time, passing the BitStream sort only seems useful when parsing
non-leaf terms. Remove them for now...