Flat parser: Add Support to Parse Variable Length Data (#157)
* Add function to parse variable length data

this function will take a BitStream at the current position and return a
variable-length data an integer as well as the bit length traversed in
order to parse the datum. It's up to the responsibility of the
caller of #getVarLenData to convert the Int to the desired plutus type.

* flat parser: parse version number of any size

Version numbers are unbounded unsigned integers. The previous
implementation only considered version numbers that fit within 7 bits.
Therefore, change the parser to scan unbounded integers.

* Add flat tests with large version numbers

These program versions require more than 7 bits to store and therefore,
each version number will exercise the function to parse variable length
data.