- Adding wrappers to crypto builtins
The Haskell backend does not have all crypto builtins avaiable. As a
matter of fact, in principle, we do not need the actual values
produced by these primitives while running the symbolic backend.
Therefore, we create function wrappers to UPLC crypto builtins to that
are bound to the crypto hooks in K when the llvm backend is used, in a
module with attribute `concrete`, and to `abstract` functions
(declared with `no-evaluators` attribute), in a module with attribute
`symbolic`.