- Changing the sort of `#eval`
In the Barcelona retreat we started to play with some proofs using the
symbolic backend. The claim we played with simply states that the
`addInteger` builtin reduces to K's `+Int` independently of the
evaluation context. Surprisingly, the prover exhibited sources of
non-determinism in rule application for term `#eval(addInteger, _, _):
since `#eval` has sort `Value`, rules that have `V:Value ~> SOME-TERM
...` would also apply leading to a number of undesirable branches.

The solution we applied in Barcelona was to change the priority of the
rule that actually simplifies `#eval` such that it takes precedence
over the other rules. There seems to be a simpler solution though: we
could simply enlarge the sort of `#eval` to one that includes
`Value`. This prevents the other rules from being applied and does not
require us to change the priority of every single rule that simplifies
`#eval`.