Add simple proofs (#162)
* draft work on simple proofs

kplc prove --directory simple-proofs/verification/haskell simple-proofs/simple-symbolic.md --debugger --debug-script k-script

* This adds basic infrastructure for doing proofs with `kplc`.

* Removed rule priority in `uplc-integer-builtins.md` to expose error message on Haskell backend.

* - Adding wrappers to crypto builtins

The Haskell backend does not have all crypto builtins avaiable. As a
matter of fact, in principle, we do not need the actual values
produced by these primitives while running the symbolic backend.
Therefore, we create function wrappers to UPLC crypto builtins to that
are bound to the crypto hooks in K when the llvm backend is used, in a
module with attribute `concrete`, and to `abstract` functions
(declared with `no-evaluators` attribute), in a module with attribute
`symbolic`.

* - Changing the sort of `#eval`

In the Barcelona retreat we started to play with some proofs using the
symbolic backend. The claim we played with simply states that the
`addInteger` builtin reduces to K's `+Int` independently of the
evaluation context. Surprisingly, the prover exhibited sources of
non-determinism in rule application for term `#eval(addInteger, _, _):
since `#eval` has sort `Value`, rules that have `V:Value ~> SOME-TERM
...` would also apply leading to a number of undesirable branches.

The solution we applied in Barcelona was to change the priority of the
rule that actually simplifies `#eval` such that it takes precedence
over the other rules. There seems to be a simpler solution though: we
could simply enlarge the sort of `#eval` to one that includes
`Value`. This prevents the other rules from being applied and does not
require us to change the priority of every single rule that simplifies
`#eval`.

* Add `function` and `functional` attributes to `#app`.

Function `#app` is total. This information is important to the
prover. This commit adds this attribute to it.

Co-authored-by: Erik Kaneda <erik.kaneda@runtimeverification.com>