test suite: compare kplc output to uplc output only
The test suite compared kplc output to uplc output contained in
.expected files as well as a previous known good kplc output stored in
.krun.expected. We made the decision to compare the output against a
previous kplc output because it allowed us to compare values of other
cells. However, we have yet to encounter a benefit for regressing on
cell values other than the K cell so remove this comparison and remove
all krun.expected files.