Move some tests from simple to new-syntax
These files contain new syntax that is yet to be adopted by uplc. Move
this to a temporary new-syntax directory. In this directory, the tests
are only run using kplc and the expected output is generated by kplc.
Once uplc adopts this syntax, these tests will be moved back to the
simple directory.