Map environment (#170)
* Use hash of a value as key to in on the heap

A more efficient environment is implemented in this PR where it is represented as a map with identifiers as keys and a stack of UPLC values as map values. This allows KPlutus to take advantage of associativity-commutativity of Map and associativity of List while updating and looking up the environment/heap.

* Reimplementing environment as a map

@bruce.collie suggested we could implement the environment as a map
from identifiers to stacks of heap indices instead of a stack of
bindings from identifiers to heap indices. This is what this commit is all about.