tests: add test cases that evaluate to (error)
Since uplc returns a non-zero status when an `error` term is encountered
during evaluation. This change adds a directory for programs that result
in this error state and a makefile target `make test-error` to execute
tests in this directory.

TODO: we also need to figure out how to incorporate uplc when running
these tests.