Integer errors (#180)
* Add tests for integer builtins

Rule coverage has shown us that there missing cases in the current
test suite. This PR adds some of these tests for integer builtins.

* Add interger builtin error tests

The acceptance criteria for milestone 1 of the KPlutus project is to
have nearly 100% rule coverage, meaning that we should make our best
efforts to get all (semantic) rules applied. This PR adds error tests
for all integer builtins essentially by either calling a given builtin
with a wrong argument type either first or second. Note that no arguments
or a single argument is not an error. It should return the resulting term.
This is case in `uplc` but not in `kplc` yet.

* No-args and one-arg tests

Move tests/error/addInteger-no-args.uplc and
tests/error/addInteger-one-arg.uplc to
tests/simple/addInteger-no-args.uplc and
tests/simple/addInteger-one-arg.uplc to

* Fix tests/failing typo and add .expected files for tests/error