Polymorphic errors (#181)
* Force and polymorphic builtins

All polymorphic builtins require an encompassing `force`. To test it
this commit makes a few changes:

1. In `uplc-syntax.md`, it organizes all builtin names into subsorts
of `BuiltinNames`. For instance, polymorphic builtin names are terms
of sort `PolyBuiltinNames` which is a subusort of `BuiltinNames`.

2. Adds a rule to `uplc-polymorphic-builtins.md` that reduces the
state to `(error)` if a polymorphic builtin does not have a `Force` in
its context.

* No-force tests

This commit adds files to `tests/error` that exercise whenever a
polymorphic builtin runs without an encompassing `force`.

* Add error tests for wrong argument type.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>