Add conditional in requires clause
Without this additional clause, the haskell backend and llvm backend
apply different rules. In this particular case, the llvm backend applies
a rule that crashes by overrunning the buffer where the haskell backend
applies rules that do not result in this condition. The divergent
behavior can be seen by running the `make test-unit-tests` command to
run the haskell backend and `make test-flat` command to run the llvm
backend.