Flat parser: add string and bytestring constants (#188)
* flat parser: add support for ByteString constants

Also add associated flat format tests and unit tests.

* flat parser: add support for string constants

Also, add tests and unit tests

* Stub out decodeBytes for different backends

At the time of writing, the `decodeBytes` hook is only implemented in
the haskell backend. For now, stub out this function and use a helper
function in `uplc-string.md`. Once the hook is implemented in the llvm
backend, remove these stubs and use decodeBytes.

* Add conditional in requires clause

Without this additional clause, the haskell backend and llvm backend
apply different rules. In this particular case, the llvm backend applies
a rule that crashes by overrunning the buffer where the haskell backend
applies rules that do not result in this condition. The divergent
behavior can be seen by running the `make test-unit-tests` command to
run the haskell backend and `make test-flat` command to run the llvm
backend.