Flat parser: enable parsing function application
Parse function application. The most notable part of this change is that
after parsing a term, the `#readProgramTerm` function returns either 1)
the term and the next bit indext to be parsed or 2) only the term
itself. The former indicates that the parser has yet to reach the end of
the input program and 2) indicates that the parse reached the end of the
program.