flat parser: add function application (#190)
* Use 7 as bit length of builtin tag

This was an error in the spec. The actual bit length is 7 rather than 8.

* Flat parser: enable parsing function application

Parse function application. The most notable part of this change is that
after parsing a term, the `#readProgramTerm` function returns either 1)
the term and the next bit indext to be parsed or 2) only the term
itself. The former indicates that the parser has yet to reach the end of
the input program and 2) indicates that the parse reached the end of the
program.

* flat parser: add simple tests

* Makefile: add flag to kompile to prevent wrapping

The error messages tend to be long. Do not wrap them and instead display long
lines for clariy.