Enable parsing bytestring inside application terms
This is enabled by applying `#resolveTerm` function to determine whether
this term is the very last term in the program. If so, return the term.
If not, return a pair containing the term as well as the next bit
position to be parsed. This change also enables all tests that use
byteString builtins as well as a few others that have been ported from
`tests/simple`