Various fixes in the flat parser and porting simple tests to flat format (#196)
* trivial: fix comment format

* Generate flat format files from tests/simple

* This commit transforms all uplc file in tests/simple into flat
  format files.

* It also add those files to tests/failing, appart from those that
  handle integer related builtins.

* add more files to tests/failing

These are failing due to the semantics returning error on partially
applied functions. Enable these tests when the semantics allows these
terms.

* ignore .hex.flat file generated from running tests

* flat parser: Retain leading 0's when reading input

Previous version of the flat parser dropped leading zeros. This change
prevents this by using the correct version of Int2Bytes which pads the
bytes instance with the apporperiate amount of 0's.

* flat parser: fix empty ByteString/String parse

Strings and ByteStrings only contain the null terminator for non-empty
string/bytestrings. Change the flat parser to account for this.

* Enable parsing bytestring inside application terms

This is enabled by applying `#resolveTerm` function to determine whether
this term is the very last term in the program. If so, return the term.
If not, return a pair containing the term as well as the next bit
position to be parsed. This change also enables all tests that use
byteString builtins as well as a few others that have been ported from
`tests/simple`

* tests/simple: add a test to exercise appending empty string

* Add expected results for all tests derived simple

* Enable force and fix rules for force/delay

This implements parsing the `force` term (which is similar to delay),
fixes an issue with precedence rules in force and delay, and enables
remaining tests that utilize polymorphic builtin functions.

Co-authored-by: Christiano Braga <christiano.braga@runtimeverification.com>