Flat parser: implement lam and var (#202)
* Pass Context to #readProgramTerm

This change prepares the parser for lambda and variable terms and results in
no functional change. The lambdaContext passed to the #readProgramTerm
function will be used by the lambda and variable terms.

* flat parser: implement lambda abstraction parser

When lambdas are parsed, a new variable name is created. The variable
name is created based off of a counter inside of the LambdaContext.

* flat tests: add simple lambda tests

These contain nested lambda term without application.

* Keep track of the next variable number during parse

The variable number is passed as TermBitLengthPair. This sort has been
renamed to `LeafTermContext`. This number is passed throughout the parse
so that each variable name is unique. After this change, terms like the
following will be parsed with different variable names.
[ ( lam v_0 ( con unit () ) ) ( lam v_1 ( con unit () ) ) ]

* add tests with lam underneath app

* Use semantic lists instead of syntactic lists

The incorrect usage of semantic casts caused the llvm backend to halt
execution. The haskell backend executes this without halting
execution (as seen by the unit tests).

* Add variable parsing

The variable is encoded as a De Bruijn indicies which is used to look up
the List of bound identifiers of the parent lambda abstraction.

* Add tests containing variable and update expected results

The expected results were generated from running uplc in the textual
format from the tests/simple directory. These test results have been
updated capturing the output of running uplc on the flat program.

* Convert uplc-examples tests to flat and generate expected values

This exercises the flat parser on larger examples and completes the flat
parser for this milestone.