Add support for typed application
Even though partial application is naturally supported by CEK
semantics, ill-typed partial applications should result in error. This
commit adds type checking (for builtin signatures) to polymorphic
builtins. We do not compute the type of a given value, we just make
sure that the signature of a given builtin is being followed. For
polymorphic builtins we check if the argument is builtin polymorphic
or fully polymorphic depending on the given builtin's signature.