Add partial application as final state (#209)
* Add partial application as final state

Partial applications are also values that should be properly
pretty-printed. They are naturally computed with by the CEK machine
but need to be transformed into values when they become final states.

* Add pretty-priting of partial application

* This commit adds pretty-printing of CEK partial application into
  UPLC partial application.

* An ill-typed partial application should result in error. This is not
  covered by this commit and will be fixed in a subsequent commit.

* Add general error rule for #eval

* This commit adds a general rule for #eval in the othwerwise case
  instead of a single rule for each builtin as before.

* Add support for typed application

Even though partial application is naturally supported by CEK
semantics, ill-typed partial applications should result in error. This
commit adds type checking (for builtin signatures) to polymorphic
builtins. We do not compute the type of a given value, we just make
sure that the signature of a given builtin is being followed. For
polymorphic builtins we check if the argument is builtin polymorphic
or fully polymorphic depending on the given builtin's signature.

* Add typechecking to Polymorphic and (some) ByteString builtins.

* Add partial application to bytestring builtins

* To be able to do proper partial application it is necessary to type
  check every argument during an application. This commit does it for
  bytestring builtins.

* Add type checking to crypto builtins

* Add partial application support to data builtins

* Add partial evaluation support to integer builtins

It also adds the tests case `tests/textual/new-syntax/mkCons-part.uplc` that
should have been added in a prior commit as well as moving all files
from `tests/new-syntax` to `tests/textual/new-syntax`

* Add support for partial application to string builtins

* Fix pretty-printing and add expected files for test cases with partial application.

* Claims about application now have to take into account that there
  can't be an application with "some ill-typed arguments". If you are
  applying to two arguments, the first necessarily has to be
  well-typed.

Co-authored-by: Christiano Braga <christiano.braga@runtimeverification.com>