Prove simple identity claim with concrete inputs (#230)
* Add identity function claim with concrete input

This simple claim aims to "iron-out" quirks in the lambda and function
application semantics.

* Add symbolic #uplcHash function

The only use of this function is to generate a unique key in a map. Use
the term `#uplcHash(Value)` itself as the key instead of getting the
actual hash value of the input.

* replace #in with in_keys(...)

`#in` is defined solely in terms of `in_keys` so replace the use with
`in_keys` which is a builtin operation.

* Modify environment functions

This commit adds various changes in the functions that operate on
environments to progress the prover.

* Add simplification rules for variables and app

Co-authored-by: rv-jenkins <admin@runtimeverification.com>