Simplify identity claim
Because `#push` is functional, we no longer need two claims for this
proof. Simplify to one claim and remove the `(v_0 in_keys(RHO))`
clauses. Due to the use of Env, we also no longer need the requires
allValuesAreList(RHO). This generalizes the claim to an arbitrary
environment and heap!