Add Env sort and make #push functional
The default Map sort for K uses `KItem`. Using KItem for value in `Map`
necessitates the use of casts when accessing map elements. This causes
functions that retrieve values from Maps to be partial and makes proof
complicated. This change implements a typed map called `Env` where the key
is a UPLC identifier and the values are `List`s. This eliminates the use
for casting to list upon retrieval and facilitates writing total
functions when working with the environment.