Clean env on return phase (#246)
* Ignore verification kompiled directory

* Drop contents of Env cell during the "return phase"

According to the CEK machine defined in the Plutus spec, the state
indicating the return phase does not contain an environment. By leaving
this cell alone, it's possible for the steps in the return phase to
maintain unneeded elements in `<Env>`. While this has no implications for
`kplc run`, this will make it difficult to write general claims. This
changes rewrite the `<Env>` cell to `.Map` when the computation
transitions between the compute phase to the return phase for constants,
variables, lambda, and delay. Builtins will be addressed in a later
commit.

* Change <Env> cell to .Map on claims that go through return phase

* Clean env cell on return phase of builtins