CBOR: implement and test decoders d and DHead
These functions are listed in section D.4 of the plutus specification
and are core functions used to parse CBOR terms. Subsequent commits
will use these functions to parse CBOR-encoded data constants and
translate them to the TextualData sort.

The UPLC-CBOR-PARSER module was imported in uplc-configuration.md to
facilitate unit testing.