Configuration: translate CBOR data to textualData
To improve readability of CBOR-encoded data constnats, convert them
to the textualData format after the input program has been parsed.
By doing so, the input program is translated to a human readable format
and the semantics of builtin functions that operate on data can
continue to use the textual format.